Wrote the logic in WindowFocuser for focusing by direction.

+++ src/external_tools/xdotool.rs
@@ -1,11 +1,11 @@
 use super::utils::get_command_output;
 
-pub fn get_current_focused_window_id() -> i32 {
+pub fn get_current_focused_window_id() -> usize {
     let output = get_command_output(&["xdotool", "getwindowfocus"])
         .trim()
         .to_owned();
 
-    output.parse::<i32>().unwrap()
+    output.parse::<usize>().unwrap()
 }
 
 #[cfg(test)]
